Fully working loop
This should have the loop just about right, and good examples of how to
do the next levels.

It's SLIGHTLY hacky, there's a fair little bit of boilerplate, BUT it
works. It doesn't quite do events, but all that'd need to do is add them
on in the same way as the next level of stars :)

There probably needs to be some polish done on the between-stars text, I
did that relatively fast.

At the moment, it pretends we've already got there, but it should be
able to handle showing progress towards the next star if that gets set (
it's just the var "progress")